requirements.txt
gensim>=4.3.2
sentence-transformers
langchain-google-genai
streamlit>=1.32.0
scikit-learn>=1.4.0
tqdm
openai
cohere
pyyaml
lancedb
python-dotenv